Incorporate review comments.
